Change the  way institution/school is handled
As institution and / school is the name of the publishing organization
it is way better to use nco:publisher directly instead of creating a dumm publisher with a affil of unknown role that has an organization
with the name of the publisher.

as nco:publisher is of type nco:Contact which is a superclass of nco:organization contact I'm still correct
with the rdf representation.

When I do the import/export from/to BibTeX I can still assign the right values as every entry that has an institution/school does not allow
a publisher name.