remove pimo:Collection concept from the program
From now on I load any pimo:Project available to Nepomuk even if it wasn't created by conquirere.

In case we get an "unknown project" we create a new default .ini file for it so the user can change the settings.
Setting files are still connected via groundingOcuurence to the project.