Add tabbed view to the documents docklet
Now you can open Documents in a tab in the background.
To do so, right click the publication in the table view and select "open in tab"

You can move around and close the tabs again. If only 1 tab is left the tabbar will be hidden.
If you select a different document the connected resource will also be opene din the sidebar.
This allows to review several document at once, without searching them again in the table list