Add cache to speedup startup of the program
You can turn on the cache in the options.
After a restart the cache will be used to fill the tables with all data.

All changes made outside of Conquirere since the last restart will be updated slowly afterwards this might lead to some wrong visula data in the table view
but the correct data is still in the nepomuk database and will be shown if you select the resource in the sidebar.
The table view will be updated like before in the background without interrupting your workflow.