Change Sparql Query/ResourceWatcher from QThread::moeToThread to QtConcurrent
In rare cases having the ResourceWatcher running in a thread lead to multiple access of the resource types and crash the program.
As threading is just needed to offer a more responsive startup it is done via QtConcurrent.

The ResourceWatcher and its connected slots for new/changed resources will be run in the main gui thread instead