Refactoring the nbibio importer/exporter/pipe classes
In order to simplify the io classes and reuse the nepomukpipe classes from the metadata extractor
all nbibio classes are refactored.

Now the main classes do not depend on kbibtex File format but rather the QVariantList approach from
the metadata extractor. In addition kbibtex <-> qvariantlist transitions are added.

In the future the kbibtex dependecy should be minimized as much as possible and only needed
for the file import/export.

Unittests are added for all classes.
Some bugs are fixed to allow flawless im- export of all publication data

newest metadata extarctor is needed
zotero sync is broken for now

in addition show plain publication (misc) entries instead of hiding them