Replace KBibTeX online search with the Metadata Extractor plugin version
Instead of KBibTeX search plugins that are written in C++ and harder to maintain
The publication websearch now uses the Metadata extractor library.

This reduce the available website a bit for now, but is a lot easier to extend via python
in the future. This also adds the ability to download references for the selected publication
while importing them.