Remove Nepomuk resource checking in the table models
Instead of a cahced Nepomuk2::Resource lookup of the updated values
the tablemode lwill now only use direct sparql queries that fetch only
the values which are displayed in the table.

Whenever the resourcewatcher sees a change in one of the resources
the lookup is done based on the resource uri, thus the bug with the
cached values in the table model that neded a restart to work is gone.